requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
Django==5.2
django-jazzmin==3.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==22.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
